pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kerf"
version = "0.1.0"
description = "Multikernel Management System"
authors = ["Cong Wang"]
readme = "README.md"
packages = [{include = "kerf", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pylibfdt = "^1.7.0"
click = "^8.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^22.0.0"
flake8 = "^5.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
kerf = "kerf.cli:main"
